[FEAT] Time is a double value now instead of long. Time unit is changed to second instead of millisecond. - Added javadoc comments to AbstractLogger.java - Add new scenario for simulating Bitcoin Network - Fixed Bitcoin Miner Node

+++ src/main/java/jabs/data/Block.java
@@ -33,6 +33,6 @@ public abstract class Block<B extends Block<B>> extends Data implements Comparab
     }
 
     public int compareTo(Block<B> b) {
-        return Long.compare(this.height, b.getHeight());
+        return Integer.compare(this.height, b.getHeight());
     }
 }
