+++ src/main/java/jabs/event/EventFactory.java
@@ -7,16 +7,16 @@ import jabs.simulator.Simulator;
 import jabs.network.Network;
 
 public class EventFactory {
-    public static void createBlockGenerationEvents(Simulator simulator, RandomnessEngine randomnessEngine, BlockchainNetwork network, int num, long timeBetweenGenerations) {
-        long time = 0;
+    public static void createBlockGenerationEvents(Simulator simulator, RandomnessEngine randomnessEngine, BlockchainNetwork network, int num, double timeBetweenGenerations) {
+        double time = 0;
         for (int i = 0; i < num; i++) {
             time += randomnessEngine.sampleExponentialDistribution(timeBetweenGenerations);
             simulator.putEvent(new BlockGenerationEvent(network.getRandomMinerByHashPower(), simulator), time);
         }
     }
 
-    public static void createTxGenerationEvents(Simulator simulator, RandomnessEngine randomnessEngine, Network network, int num, long timeBetweenGenerations) {
-        long time = 0;
+    public static void createTxGenerationEvents(Simulator simulator, RandomnessEngine randomnessEngine, Network network, int num, double timeBetweenGenerations) {
+        double time = 0;
         for (int i = 0; i < num; i++) {
             time += randomnessEngine.sampleExponentialDistribution(timeBetweenGenerations);
             simulator.putEvent(new TxGenerationEvent(network.getRandomNode()), time);
