+++ src/main/java/jabs/event/TxGeneratorProcessFixedNode.java
@@ -5,7 +5,7 @@ import jabs.randengine.RandomnessEngine;
 import jabs.simulator.Simulator;
 
 public class TxGeneratorProcessFixedNode extends AbstractTxGeneratorProcess {
-    public TxGeneratorProcessFixedNode(Simulator simulator, RandomnessEngine randomnessEngine, Node node, long averageTimeBetweenTxs) {
+    public TxGeneratorProcessFixedNode(Simulator simulator, RandomnessEngine randomnessEngine, Node node, double averageTimeBetweenTxs) {
         super(simulator, randomnessEngine, averageTimeBetweenTxs);
         this.node = node;
     }
